Refactoring the review site.
