+++ libs/application/templates/parental-leave/src/fields/Review/review-groups/props.ts
@@ -1,0 +1,9 @@
+import { Application, FormValue } from "@island.is/application/types"
+
+export interface ReviewGroupProps {
+  application: Application<FormValue>
+  editable: boolean
+  groupHasNoErrors: (ids: string[]) => boolean
+  hasError: (id: string) => string
+  goToScreen?: (id: string) => void
+};
